Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long int s;
 	long long int a;
 
-	printf("size of a char: %lu byte(s)\n", sizeof(f));
-	printf("size o int: %lu byte(s)\n", sizeof(j));
-	printf("size of float: %lu byte(s)\n", sizeof(k));
-	printf("size of long int: %lu byte(s)\n", sizeof(s));
-	printf("size of long long int: %lu byte(s)\n", sizeof(a));
+	printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(f));
+	printf("size o int: %lu byte(s)\n", (unsigned long)sizeof(j));
+	printf("size of float: %lu byte(s)\n", (unsigned long)sizeof(k));
+	printf("size of long int: %lu byte(s)\n", (unsigned long)sizeof(s));
+	printf("size of long long int: %lu byte(s)\n", (unsigned long)sizeof(a));
 	return (0);
 }
